requirements.txt
argparse
arrow
filetype
nextcloud-api-wrapper
python-telegram-bot
toml
